pyproject.toml
[project]
name = "podcast"
version = "0.1.0"
description = "Private podcast feed generator"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "pydantic-settings>=2.7",
    "jinja2>=3.1",
    "python-multipart>=0.0.18",
    "anthropic>=0.52",
    "feedgen>=1.0",
    "pydub>=0.25",
    "chatterbox-tts>=0.1",
    "torchaudio>=2.0",
    "httpx>=0.28",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
